requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pandas
numpy
python-dotenv
pydantic
google-generativeai
faiss-cpu
requests
streamlit
pytest
httpx
loguru
matplotlib
tqdm
sentence-transformers
groq